requirements.txt
langchain-google-genai
langchain-groq
langchain-cohere
langchain-openai
langchain-community
PyPDF2  
faiss-cpu
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
python-dotenv
fastapi
uvicorn[standard]
slackclient
slack_bolt
slackeventsapi
msal
pymupdf4llm
matplotlib
umap-learn
pandas
numpy
boto3
semantic-router
sqlalchemy